Data.SortedList

Description

This module defines a type for sorted lists, together with several functions to create and use values of that type. Many operations are optimized to take advantage of the list being sorted.

Type

data SortedList a

Type of sorted lists. Any (non-bottom) value of this type is a sorted list. Use the Monoid instance to merge sorted lists.

List conversions

toSortedList :: Ord a => [a] -> SortedList a

Create a SortedList by sorting a regular list.

fromSortedList :: SortedList a -> [a]

O(1). Create a list from a SortedList. The returned list is guaranteed to be sorted.

Construction

singleton :: a -> SortedList a

O(1). Create a sorted list with only one element.

repeat :: a -> SortedList a

An infinite list with all its elements equal to the given argument.

replicate :: Int -> a -> SortedList a

Replicate a given number of times a single element.

iterate :: Ord a => (a -> a) -> a -> SortedList a

Create a sorted list by repeatedly applying the same function to an element, until the image by that function is stricly less than its argument. In other words:

iterate f x = [x, f x, f (f x), ... ]

With the list ending whenever f (f (... (f (f x)) ...)) < f (... (f (f x)) ...). If this never happens, the list will be infinite.

By definition:

iterate f = unfoldr (\x -> Just (x, f x))

Deconstruction

uncons :: SortedList a -> Maybe (a, SortedList a)

O(1). Decompose a sorted list into its minimal element and the rest. If the list is empty, it returns Nothing.

Inserting

insert :: Ord a => a -> SortedList a -> SortedList a

O(n). Insert a new element in a sorted list.

Deleting

delete :: Ord a => a -> SortedList a -> SortedList a

Delete the first occurrence of the given element.

deleteAll :: Ord a => a -> SortedList a -> SortedList a

Delete all occurrences of the given element.

Sublists

take :: Int -> SortedList a -> SortedList a

Extract the prefix with the given length from a sorted list.

drop :: Int -> SortedList a -> SortedList a

Drop the given number of elements from a sorted list, starting from the smallest and following ascending order.

splitAt :: Int -> SortedList a -> (SortedList a, SortedList a)

Split a sorted list in two sublists, with the first one having length equal to the given argument, except when the length of the list is less than that.

takeWhile :: (a -> Bool) -> SortedList a -> SortedList a

Return the longest prefix of a sorted list of elements that satisfy the given condition.

dropWhile :: (a -> Bool) -> SortedList a -> SortedList a

Return the suffix remaining after dropping the longest prefix of elements that satisfy the given condition.

span :: (a -> Bool) -> SortedList a -> (SortedList a, SortedList a)

Return the longest prefix of a sorted list of elements that satisfy the given condition, and the rest of the list.

Filtering

partition :: (a -> Bool) -> SortedList a -> (SortedList a, SortedList a)

O(n). Divide a sorted list into two lists, one with all the elements that satisfy the given predicate, and another list with the rest of elements.

filter :: (a -> Bool) -> SortedList a -> SortedList a

O(n). Extract the elements of a list that satisfy the predicate.

filterLT :: Ord a => a -> SortedList a -> SortedList a

O(n). Select only elements that are strictly less than the argument.

filterGT :: Ord a => a -> SortedList a -> SortedList a

O(n). Select only elements that are strictly greater than the argument.

filterLE :: Ord a => a -> SortedList a -> SortedList a

O(n). Select only elements less or equal to the argument.

filterGE :: Ord a => a -> SortedList a -> SortedList a

O(n). Select only elements greater or equal to the argument.

Queries

elemOrd :: Ord a => a -> SortedList a -> Bool

O(n). An efficient implementation of elem, using the Ord instance of the elements in a sorted list. It only traverses the whole list if the requested element is greater than all the elements in the sorted list.

findIndices :: (a -> Bool) -> SortedList a -> SortedList Int

O(n). Return the indices of all elements in a sorted list that satisfy the given condition.

map function

map :: Ord b => (a -> b) -> SortedList a -> SortedList b

Map a function over all the elements of a sorted list. Note that map will hang if the argument is an infinite list.

Even though SortedList can't be made an instance of Functor, map does hold the Functor laws (for finite lists). We can't however write an instance because of the Ord instance requirement on the type of the elements of the result list. Therefore, while SortedList is not a functor type in general, it is when restricted to elements of orderable types (for finite lists).

The complexity range goes from O(n) (if the function is monotonically increasing) to O(n²) (if the function is monotonically decreasing). These are the best and worst case scenarios. We provide an alternative (mapDec) where monotonically decreasing functions are the best case scenario.

mapDec :: Ord b => (a -> b) -> SortedList a -> SortedList b

Just like map, but favoring functions that are monotonically decreasing instead of those that are monotonically increasing.

Unfolding

unfoldr :: Ord a => (b -> Maybe (a, b)) -> b -> SortedList a

Dual (sort of) to foldr for sorted lists. It builds a sorted list from a generator function and an initial element. The generator function is applied to the initial element, and then it will produce either Nothing - meaning that the list building must stop - or Just applied to the value that is going to be added to the list, and a new accumulator to be fed to the generator function. The list building will stop prematurely if the generator function happens to create an element for the list that is strictly smaller than the previous value.

Others

reverse :: SortedList a -> SortedList (Down a)

O(n). Reverse a sorted list. The result uses Down, thus it is a sorted list as well. The following equality holds for any sorted list xs:

map Down xs = reverse xs

Only available from base version 4.6.0.0.

reverseDown :: SortedList (Down a) -> SortedList a

O(n). Reverse a sorted list with elements embedded in the Down type.

Only available from base version 4.6.0.0.

Set operations

nub :: Eq a => SortedList a -> SortedList a

O(n). Remove duplicate elements from a sorted list.

intersect :: Ord a => SortedList a -> SortedList a -> SortedList a

O(n+m). Intersection of sorted lists. If the first list contains duplicates, so will the result.

union :: Ord a => SortedList a -> SortedList a -> SortedList a

Union of sorted lists. Duplicates, and elements of the first list, are removed from the the second list, but if the first list contains duplicates, so will the result.